[IR] Don't pad argument list with nulls when changing callee
When reassigning IrMemberAccessExpression.symbol to a function that has
more parameters than before, new nulls were appended to the arguments
list. This was only done to maintain a correct old<->new parameter API
bridge. It is quite fragile and potentially unexpected mechanism, so
it's to better remove it and leave argument list for manual patches.

KT-76934