[Wasm/JS] allow JsReference<C> "as C"/"is C" operations
As JsReference<C> can be implicitly converted to C (e.g. by assignment
to a variable of `Any` type), the type (`is`/`as`) and identity (`===`)
operations between JsReference<C> and Kotlin type `K` shall be allowed,
if `C` can potentially be cast to `K`.
Also, any `JsAny` reference can contain `JsReference`, so operations
between JsAny and Kotlin types shall be allowed as well.

Co-authored-by: Nikolay Lunyak
- move the checker components to `:checkers`
- Simplify the main check of `FirWasmJsEqualityChecker`
- Rewrite the flow of `FirWasmJsEqualityChecker`

^KT-81637 Fixed