[FIR] Add checks if the evaluator can interpret a given function.
Previously we used `IntrinsicConstEvaluation` annotation as a
marker that a function can be evaluated. If we mark more functions,
then we will have a situation when compiler can't consume stdlib
of plus one version. This happens because IrInterpreter is trying
to evaluate such functions, but it can't find them in a special
generated map.

To overcome this, we introduced a new frontend-only map that tells
if a function can be evaluated. This map will be removed in the
future.

^KT-80804