FIR: Force Unit-coercion when ELA is run with a Unit-constraint TV
For non-ELA case, it should not really matter, because for a single
candidate we would set return type ref to Unit if 
`willEventuallyBecomeUnit == true` at `analyzeLambda`, 
but for ELA we need to check it manually.

NB: Due to the explanation above, it should not affect non-ELA case,
so no LanguageFeature barrier is used.